+++ selfdrive/car/docs_definitions.py
@@ -63,6 +63,7 @@ class CarInfo:
     if self.min_enable_speed is not None:
       min_enable_speed = self.min_enable_speed
 
+    self.car_name = CP.carName
     self.make, self.model = self.name.split(' ', 1)
     self.row = {
       Column.MAKE: self.make,
@@ -75,6 +76,10 @@ class CarInfo:
       Column.STEERING_TORQUE: self.good_torque,
       Column.MAINTAINED: CP.carFingerprint not in non_tested_cars,
     }
+
+    if CP.notCar:
+      for col in StarColumns:
+        self.row[col] = True
 
     self.all_footnotes = all_footnotes
     for column in StarColumns: